pyproject.toml
[project]
name = "robothector"
version = "0.1.0"
description = "Remote-controlled robot with video streaming"
requires-python = ">=3.11"

[project.optional-dependencies]
server = [
    "flask>=2.2",
    "websockets>=12.0",
    "picamera2>=0.3; platform_machine=='aarch64'",
    "RPi.GPIO>=0.7; platform_machine=='aarch64'",
    "pygame>=2.1",
]
client = [
    "pygame>=2.1",
    "websocket-client>=1.6",
]

[tool.uv]
dev-dependencies = []
